fix filter layer

+++ metrics-util/src/layers/filter.rs
@@ -55,7 +55,7 @@ impl<R: Recorder> Recorder for Filter<R> {
         self.inner.update_gauge(key, value);
     }
 
-    fn record_histogram(&self, key: Key, value: u64) {
+    fn record_histogram(&self, key: Key, value: f64) {
         if self.should_filter(&key) {
             return;
         }
